#5 - 쿼리 성능 개선 & record 문법
